+++ R/install-github.R
@@ -41,33 +41,27 @@
 #' install_github("hadley/private", auth_token = "abc")
 #'
 #' }
-install_github <- function(repo, username = NULL,
+install_github <- function(repo,
                            ref = "master", subdir = NULL,
                            auth_token = github_pat(),
                            host = "api.github.com", ...) {
 
-  remotes <- lapply(repo, github_remote, username = username, ref = ref,
+  remotes <- lapply(repo, github_remote, ref = ref,
     subdir = subdir, auth_token = auth_token, host = host)
 
   install_remotes(remotes, ...)
 }
 
-github_remote <- function(repo, username = NULL, ref = NULL, subdir = NULL,
+github_remote <- function(repo, ref = NULL, subdir = NULL,
                        auth_token = github_pat(), sha = NULL,
                        host = "api.github.com") {
 
   meta <- parse_git_repo(repo)
   meta <- github_resolve_ref(meta$ref %||% ref, meta, auth_token)
 
-  if (is.null(meta$username)) {
-    meta$username <- username %||% getOption("github.user") %||%
-      stop("Unknown username.")
-    warning("Username parameter is deprecated. Please use ",
-      username, "/", repo, call. = FALSE)
-  }
-
   remote("github",
     host = host,
+    package = meta$package,
     repo = meta$repo,
     subdir = meta$subdir %||% subdir,
     username = meta$username,
@@ -129,6 +123,7 @@ remote_metadata.github_remote <- function(x, bundle = NULL, source = NULL) {
   list(
     RemoteType = "github",
     RemoteHost = x$host,
+    RemotePackage = x$package,
     RemoteRepo = x$repo,
     RemoteUsername = x$username,
     RemoteRef = x$ref,
@@ -213,127 +208,26 @@ github_resolve_ref.github_release <- function(x, params, ..., auth_token = NULL)
   params
 }
 
-#' Parse a remote git repo specification
-#'
-#' A remote repo can be specified in two ways:
-#' \describe{
-#' \item{as a URL}{\code{parse_github_url()} handles HTTPS and SSH remote URLs
-#' and various GitHub browser URLs}
-#' \item{via a shorthand}{\code{parse_repo_spec()} handles this concise form:
-#' \code{[username/]repo[/subdir][#pull|@ref|@*release]}}
-#' }
-#'
-#' @param repo Character scalar, the repo specification.
-#' @return List with members: \code{username}, \code{repo}, \code{subdir}
-#'   \code{ref}, \code{pull}, \code{release}, some which will be empty.
-#'
-#' @name parse-git-repo
-#' @examples
-#' parse_repo_spec("metacran/crandb")
-#' parse_repo_spec("jimhester/covr#47")        ## pull request
-#' parse_repo_spec("jeroen/curl@v0.9.3")       ## specific tag
-#' parse_repo_spec("tidyverse/dplyr@*release") ## shorthand for latest release
-#' parse_repo_spec("r-lib/remotes@550a3c7d3f9e1493a2ba") ## commit SHA
-#'
-#' parse_github_url("https://github.com/jeroen/curl.git")
-#' parse_github_url("git@github.com:metacran/crandb.git")
-#' parse_github_url("https://github.com/jimhester/covr")
-#' parse_github_url("https://github.example.com/user/repo.git")
-#' parse_github_url("git@github.example.com:user/repo.git")
-#'
-#' parse_github_url("https://github.com/r-lib/remotes/pull/108")
-#' parse_github_url("https://github.com/r-lib/remotes/tree/name-of-branch")
-#' parse_github_url("https://github.com/r-lib/remotes/commit/1234567")
-#' parse_github_url("https://github.com/r-lib/remotes/releases/latest")
-#' parse_github_url("https://github.com/r-lib/remotes/releases/tag/1.0.0")
-NULL
-
 #' @export
-#' @rdname parse-git-repo
-parse_repo_spec <- function(repo) {
-  username_rx <- "(?:(?<username>[^/]+)/)?"
-  repo_rx     <- "(?<repo>[^/@#]+)"
-  subdir_rx   <- "(?:/(?<subdir>[^@#]*[^@#/])/?)?"
-  ref_rx      <- "(?:@(?<ref>[^*].*))"
-  pull_rx     <- "(?:#(?<pull>[0-9]+))"
-  release_rx  <- "(?:@(?<release>[*]release))"
-  ref_or_pull_or_release_rx <- sprintf(
-    "(?:%s|%s|%s)?", ref_rx, pull_rx, release_rx
-  )
-  spec_rx  <- sprintf(
-    "^%s%s%s%s$", username_rx, repo_rx, subdir_rx, ref_or_pull_or_release_rx
-  )
-  params <- as.list(re_match(text = repo, pattern = spec_rx))
+remote_package_name.github_remote <- function(remote, ..., use_local = TRUE,
+  use_curl = is_installed("curl")) {
 
-  if (is.na(params$.match)) {
-    stop(sprintf("Invalid git repo specification: '%s'", repo))
+  # If the package name was explicitly specified, use that
+  if (!is.null(remote$package)) {
+    return(remote$package)
   }
 
-  params[grepl("^[^\\.]", names(params))]
-}
-
-#' @export
-#' @rdname parse-git-repo
-parse_github_repo_spec <- parse_repo_spec
-
-#' @export
-#' @rdname parse-git-repo
-parse_github_url <- function(repo) {
-  prefix_rx <- "(?:github[^/:]+[/:])"
-  username_rx <- "(?:(?<username>[^/]+)/)"
-  repo_rx     <- "(?<repo>[^/@#]+)"
-  ref_rx <- "(?:(?:tree|commit|releases/tag)/(?<ref>.+$))"
-  pull_rx <- "(?:pull/(?<pull>.+$))"
-  release_rx <- "(?:releases/)(?<release>.+$)"
-  ref_or_pull_or_release_rx <- sprintf(
-    "(?:/(%s|%s|%s))?", ref_rx, pull_rx, release_rx
-  )
-  url_rx  <- sprintf(
-    "%s%s%s%s",
-    prefix_rx, username_rx, repo_rx, ref_or_pull_or_release_rx
-  )
-  params <- as.list(re_match(text = repo, pattern = url_rx))
-
-  if (is.na(params$.match)) {
-    stop(sprintf("Invalid GitHub URL: '%s'", repo))
-  }
-  if (params$ref == "" && params$pull == "" && params$release == "") {
-    params$repo <- gsub("\\.git$", "", params$repo)
-  }
-  if (params$release == "latest") {
-    params$release <- "*release"
+  # Otherwise if the repo is an already installed package assume that.
+  if (isTRUE(use_local)) {
+    local_name <- suppressWarnings(utils::packageDescription(remote$repo, fields = "Package"))
+    if (!is.na(local_name)) {
+      return(local_name)
+    }
   }
 
-  params[grepl("^[^\\.]", names(params))]
-}
-
-parse_git_repo <- function(repo) {
-
-  if (grepl("^https://github|^git@github", repo)) {
-    params <- parse_github_url(repo)
-  } else {
-    params <- parse_repo_spec(repo)
-  }
-  params <- params[viapply(params, nchar) > 0]
-
-  if (!is.null(params$pull)) {
-    params$ref <- github_pull(params$pull)
-    params$pull <- NULL
-  }
-
-  if (!is.null(params$release)) {
-    params$ref <- github_release()
-    params$release <- NULL
-  }
-
-  params
-}
-
-#' @export
-remote_package_name.github_remote <- function(remote, ...) {
-
+  # Otherwise lookup the package name from the remote DESCRIPTION file
   desc <- github_DESCRIPTION(username = remote$username, repo = remote$repo,
-    host = remote$host, ref = remote$ref, pat = remote$auth_token)
+    host = remote$host, ref = remote$ref, pat = remote$auth_token, use_curl = use_curl)
 
   if (is.null(desc)) {
     return(NA_character_)
@@ -347,9 +241,9 @@ remote_package_name.github_remote <- function(remote, ...) {
 }
 
 #' @export
-remote_sha.github_remote <- function(remote, ...) {
+remote_sha.github_remote <- function(remote, ..., use_curl = is_installed("curl")) {
   github_commit(username = remote$username, repo = remote$repo,
-    host = remote$host, ref = remote$ref, pat = remote$auth_token)
+    host = remote$host, ref = remote$ref, pat = remote$auth_token, use_curl = use_curl)
 }
 
 #' @export